hw3/booking-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
